Detect the presence/absence of a match

Vectorised over string and pattern. Actually equivalent to grepl(pattern, x) as returns FALSE for NAs (unlike stringr::str_detect()). This behavior is useful when searching comments many of which are NA to indicate no comments present.

Usage

str_detect2(string, pattern, negate = FALSE)

Arguments

string

Input vector. Either a character vector, or something coercible to one.

pattern

Pattern to look for.

The default interpretation is a regular expression, as described in vignette("regular-expressions"). Use regex() for finer control of the matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using fixed(). This is fast, but approximate. Generally, for matching human text, you'll want coll() which respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with boundary(). An empty pattern, "", is equivalent to boundary("character").

negate

If TRUE, inverts the resulting boolean vector.

Value

A logical vector the same length as string/pattern.

See also

Examples

x <- c("b", NA, "ab")
pattern <- "^a"
grepl(pattern, x)
#> [1] FALSE FALSE  TRUE
stringr::str_detect(x, pattern)
#> [1] FALSE    NA  TRUE
str_detect2(x, pattern)
#> [1] FALSE FALSE  TRUE